feat(server): show port server is running on
Have you ever started up a server and not been certain what port the
server is listening on? Me too. This is a pretty basic thing we can add
in that makes it easier for a server admin to tell what is going on with
their server during startup.

DCO-1.1-Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>